requirements.txt
gym==0.19.0
matplotlib==3.3.1
numba==0.57.1
numpy==1.22.0
Pillow==10.0.0
pyglet==1.4.11
PyOpenGL==3.1.7
PyYAML==6.0
scipy==1.10.1
torch==1.13.1
torchvision==0.14.1
torchviz==0.0.2
Werkzeug==2.2.2
